move vm.ts into vm/
