+++ benchmarks/vm/memory.benchmark.ts
@@ -2,7 +2,7 @@ import { PluginEvent } from '@posthog/plugin-scaffold/src/types'
 
 import { createServer } from '../../src/server'
 import { Plugin, PluginConfig, PluginConfigVMReponse } from '../../src/types'
-import { createPluginConfigVM } from '../../src/vm'
+import { createPluginConfigVM } from '../../src/vm/vm'
 import { commonOrganizationId } from '../../tests/helpers/plugins'
 
 jest.mock('../../src/sql')
