+++ src/vm/extensions/console.ts
