+++ src/vm/extensions/storage.ts
@@ -1,6 +1,6 @@
 import { StorageExtension } from '@posthog/plugin-scaffold'
 
-import { PluginConfig, PluginsServer } from '../types'
+import { PluginConfig, PluginsServer } from '../../types'
 
 export function createStorage(server: PluginsServer, pluginConfig: PluginConfig): StorageExtension {
     const get = async function (key: string, defaultValue: unknown): Promise<unknown> {
